move disk/net functions in their own C files
